ui is working nice
