+++ frontend/ui/src/pages/Channels/Providers/Airy/ChatPlugin/sections/ConnectNewChatPlugin.module.scss
@@ -1,0 +1,17 @@
+.newPageParagraph {
+  margin: 24px 0;
+  color: var(--color-text-gray);
+}
+
+.formWrapper {
+  width: 340px;
+  align-self: center;
+}
+
+.formRow {
+  margin-bottom: 16px;
+}
+
+.settings {
+  width: 20rem;
+}
