+++ frontend/ui/src/pages/Channels/Providers/Airy/ChatPlugin/sections/EditChatPlugin.module.scss
@@ -1,0 +1,40 @@
+.formWrapper {
+  width: 340px;
+  align-self: center;
+  display: flex;
+}
+
+.settings {
+  width: 20rem;
+}
+
+.updatePageParagraph {
+  margin: 24px 0;
+  color: var(--color-text-gray);
+}
+
+.tabView {
+  display: flex;
+  margin-bottom: 16px;
+  list-style-type: none;
+
+  li {
+    margin-right: 16px;
+  }
+}
+
+.tabEntrySelected {
+  border-bottom: 2px solid var(--color-airy-blue);
+  a {
+    color: var(--color-text-contrast);
+    font-weight: bold;
+    text-decoration: none;
+  }
+}
+
+.tabEntry {
+  a {
+    color: var(--color-airy-blue);
+    text-decoration: none;
+  }
+}
